# backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API - Sistema de Gestão de Cursos e Anotações",
      "version": "1.0.0",
      "description": "Documentação da API com Swagger"
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      }
    ],
    "paths": {
      "/api/auth/register": {
        "post": {
          "tags": ["Autenticação"],
          "summary": "Registrar novo usuário",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterUser"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Usuário criado com sucesso"
            }
          }
        }
      },
      "/api/auth/login": {
        "post": {
          "tags": ["Autenticação"],
          "summary": "Login do usuário",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginUser"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Login realizado com sucesso"
            }
          }
        }
      },
      "/api/courses": {
        "get": {
          "tags": ["Cursos"],
          "summary": "Listar cursos",
          "responses": {
            "200": {
              "description": "Lista de cursos"
            }
          }
        },
        "post": {
          "tags": ["Cursos"],
          "summary": "Criar novo curso",
          "security": [{ "bearerAuth": [] }],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Curso criado"
            }
          }
        }
      },
      "/api/notes": {
        "get": {
          "tags": ["Notas"],
          "summary": "Listar anotações do usuário",
          "security": [{ "bearerAuth": [] }],
          "responses": {
            "200": {
              "description": "Lista de anotações",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": { "$ref": "#/components/schemas/Note" }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": ["Notas"],
          "summary": "Criar nova anotação",
          "security": [{ "bearerAuth": [] }],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewNote"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Anotação criada com sucesso",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Note"
                  }
                }
              }
            }
          }
        }
      },
      "/api/notes/{id}": {
        "delete": {
          "tags": ["Notas"],
          "summary": "Excluir anotação",
          "security": [{ "bearerAuth": [] }],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": { "type": "string" }
            }
          ],
          "responses": {
            "200": { "description": "Anotação excluída" },
            "404": { "description": "Não encontrada" }
          }
        }
      }
    },
    "components": {
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      },
      "schemas": {
        "RegisterUser": {
          "type": "object",
          "required": ["name", "email", "password"],
          "properties": {
            "name": { "type": "string" },
            "email": { "type": "string", "format": "email" },
            "password": { "type": "string" },
            "role": { "type": "string", "enum": ["student", "admin"] }
          }
        },
        "LoginUser": {
          "type": "object",
          "required": ["email", "password"],
          "properties": {
            "email": { "type": "string", "format": "email" },
            "password": { "type": "string" }
          }
        },
        "Course": {
          "type": "object",
          "required": ["title", "description"],
          "properties": {
            "title": { "type": "string" },
            "description": { "type": "string" }
          }
        },
        "Note": {
          "type": "object",
          "properties": {
            "_id": { "type": "string" },
            "userId": { "type": "string" },
            "courseId": { "type": "number" },
            "content": { "type": "string" },
            "createdAt": { "type": "string", "format": "date-time" }
          }
        },
        "NewNote": {
          "type": "object",
          "required": ["courseId", "content"],
          "properties": {
            "courseId": { "type": "number" },
            "content": { "type": "string" }
          }
        }
      }
    }
  }
  